refactor: consumer 패키지 하위 구조를 도메인별로 세분화
